+++ backend/src/routers/v1/users/index.ts
@@ -5,7 +5,6 @@ import * as passport from 'passport';
 import {
   PostService,
   UserService,
-  GitService,
   BlogService,
   TistoryService,
   NotifyService,
@@ -119,19 +118,6 @@ export default class UsersRouter {
     return response.json({ code: RESPONSECODE.SUCCESS, data: posts });
   }
 
-  @Get('/:userID/repositories')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async getRepoList(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const githubUsername = await UserService.findGithubUsernameForUserID(userID);
-
-    const result = await GitService.findRepoList(githubUsername as string);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
   @Get('/:userID/notifies')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async getNotifyList(@Req() request: Request, @Res() response: Response) {
@@ -149,36 +135,6 @@ export default class UsersRouter {
     }
     const postContent = await TistoryService.findPostContent(userID, identity, postID);
     return response.json({ code: RESPONSECODE.SUCCESS, data: postContent });
-  }
-
-  @Get('/:userID/repositories/:repoName')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async getRepo(@Req() request: Request, @Res() response: Response) {
-    const { userID, repoName } = request.params;
-    if (userID !== request.user!.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const githubUsername = await UserService.findUserGithubUsername(userID);
-    if (githubUsername === undefined) {
-      throw new Error(ERROR.INVALID_GITHUB_USERNAME);
-    }
-    const result = await GitService.findRepo(githubUsername, repoName);
-    return response.json({
-      code: RESPONSECODE.SUCCESS,
-      data: {
-        title: result.repoName,
-        type: 'repository',
-        content: result.content,
-        link: result.repoUrl,
-        info: {
-          starCount: result.starCount,
-          forkCount: result.forkCount,
-          language: result.languageInfo,
-        },
-        identity: githubUsername,
-        target: repoName,
-      },
-    });
   }
 
   @Put('/:userID/settings')
